refactor for better rendering
